requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
drf-spectacular>=0.26.1,<0.27
pandas>=1.5.3,<1.6
django-extensions>=3.2.1,<3.3